tests/wpt/web-platform-tests/app-history/navigate/return-value/navigate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<script>
promise_test(async t => {
  const result = appHistory.navigate("#1");
  await assertBothFulfill(t, result, appHistory.current);
  assert_equals((new URL(appHistory.current.url)).hash, "#1");
}, "navigate() promises");
</script>
